Update typedText.jsx
Do we need a new typeString at line 331 to separate the quest title/objective from the story text? I don't want to mess with the code here, but it's not the same as the other sections.

I did some changes starting at line 597, because it seems like the wrong class is assigned to the dialogue.